fix(syslog): 修复字段拼写及完善日志展示

+++ vea/build.sh
@@ -2,4 +2,7 @@
 
 # npm config set registry http://mirrors.cloud.tencent.com/npm/
 pnpm run build
-cp src/assets/favicon.ico ./dist/
+# 如果dist/favicon.ico存在则复制到dist/
+if [ -f "src/assets/favicon.ico" ]; then
+  cp src/assets/favicon.ico ./dist/
+fi
